requirements.txt
absl-py==2.1.0
actionlib==1.14.0
angles==1.9.13
anyio @ file:///tmp/build/80754af9/anyio_1644481697509/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569681205/work
attrs @ file:///croot/attrs_1668696182826/work
Babel @ file:///croot/babel_1671781930836/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
brotlipy==0.7.0
cachetools==5.5.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle==2.2.1
coloredlogs==15.0.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography @ file:///croot/cryptography_1677533068310/work
cv-bridge==1.16.2
cycler==0.11.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
flatbuffers==24.3.25
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.34.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
gym-notices==0.0.8
humanfriendly==10.0
idna==3.10
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///tmp/abs_94gruux8u8/croots/recipe/ipython_1659529858706/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///croot/ipywidgets_1701289330913/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///croot/jsonschema_1676558650973/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console @ file:///croot/jupyter_console_1671541909316/work
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyter_client @ file:///croot/jupyter_client_1676329080601/work
jupyter_core @ file:///croot/jupyter_core_1668084443243/work
jupyterlab @ file:///croot/jupyterlab_1675354114448/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
jupyterlab_server @ file:///croot/jupyterlab_server_1677143054853/work
kiwisolver==1.4.5
llvmlite==0.39.1
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
message-filters==1.16.0
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mpmath==1.3.0
nb-conda==2.2.1
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1606775986548/work
nbclassic @ file:///croot/nbclassic_1676902904845/work
nbclient @ file:///tmp/build/80754af9/nbclient_1650308370114/work
nbconvert @ file:///croot/nbconvert_1668450669124/work
nbformat @ file:///croot/nbformat_1670352325207/work
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==2.6.3
noise==1.2.2
notebook @ file:///croot/notebook_1668179881751/work
notebook_shim @ file:///croot/notebook-shim_1668160579331/work
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.14.1
opencv-python==4.10.0.84
packaging==24.0
pandas==1.3.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pkgutil_resolve_name @ file:///opt/conda/conda-bld/pkgutil-resolve-name_1661463321198/work
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pynndescent==0.5.13
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.4
pyrealsense2==2.55.1.6486
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz @ file:///croot/pytz_1671697431263/work
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qtconsole @ file:///croot/qtconsole_1674008428467/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
requests==2.31.0
requests-oauthlib==2.0.0
rosbag==1.16.0
rosclean==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslz4==1.16.0
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sensor-msgs==1.13.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smclib==1.8.6
sniffio @ file:///tmp/build/80754af9/sniffio_1614030478629/work
soupsieve @ file:///croot/soupsieve_1666296392845/work
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado @ file:///croot/terminado_1671751832461/work
tf==1.13.2
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2 @ file:///croot/tinycss2_1668168815555/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
topic-tools==1.16.0
torch==1.12.1
torchaudio==0.12.1
torchcam==0.3.2
torchvision==0.13.1
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.66.5
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions==4.7.1
umap-learn==0.5.6
urllib3==1.26.20
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///tmp/build/80754af9/websocket-client_1614804260725/work
Werkzeug==2.2.3
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354203/work
xacro==1.14.18
zipp==3.15.0
